Move RateLimiter to it's proper spot in the network package (vs utils)
The implementation of RateLimiter is too simplistic and usage-specific to really belong in utils, so it should be moved to network where it belongs.
